tec: Extract the dao save method
