+++ src/models/dao/Token.php
@@ -8,6 +8,8 @@ namespace flusio\models\dao;
  */
 class Token extends \Minz\DatabaseModel
 {
+    use SaveHelper;
+
     /**
      * @throws \Minz\Errors\DatabaseError
      */
@@ -15,26 +17,5 @@ class Token extends \Minz\DatabaseModel
     {
         $properties = array_keys(\flusio\models\Token::PROPERTIES);
         parent::__construct('tokens', 'token', $properties);
-    }
-
-    /**
-     * Create or update a model in database
-     *
-     * @param \Minz\Model $model
-     *
-     * @throws \Minz\DatabaseModelError
-     *
-     * @return string The primary key value of the model
-     */
-    public function save($model)
-    {
-        $values = $model->toValues();
-        if ($model->created_at) {
-            $this->update($model->token, $values);
-            return $model->token;
-        } else {
-            $values['created_at'] = \Minz\Time::now()->format(\Minz\Model::DATETIME_FORMAT);
-            return $this->create($values);
-        }
     }
 }
