+++ src/models/dao/User.php
@@ -10,6 +10,8 @@ namespace flusio\models\dao;
  */
 class User extends \Minz\DatabaseModel
 {
+    use SaveHelper;
+
     /**
      * @throws \Minz\Errors\DatabaseError
      */
@@ -17,27 +19,6 @@ class User extends \Minz\DatabaseModel
     {
         $properties = array_keys(\flusio\models\User::PROPERTIES);
         parent::__construct('users', 'id', $properties);
-    }
-
-    /**
-     * Create or update a model in database
-     *
-     * @param \Minz\Model $model
-     *
-     * @throws \Minz\DatabaseModelError
-     *
-     * @return string The primary key value of the model
-     */
-    public function save($model)
-    {
-        $values = $model->toValues();
-        if ($model->created_at) {
-            $this->update($model->id, $values);
-            return $model->id;
-        } else {
-            $values['created_at'] = \Minz\Time::now()->format(\Minz\Model::DATETIME_FORMAT);
-            return $this->create($values);
-        }
     }
 
     /**
